+++ apiserver/test/e2e/cluster_server_e2e_test.go
@@ -1,20 +1,16 @@
 package e2e
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"testing"
-	"time"
 
 	kuberayHTTP "github.com/ray-project/kuberay/apiserver/pkg/http"
 	api "github.com/ray-project/kuberay/proto/go_client"
 
 	rayv1api "github.com/ray-project/kuberay/ray-operator/apis/ray/v1"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"k8s.io/apimachinery/pkg/util/wait"
 )
 
 // TestCreateClusterEndpoint sequentially iterates over the create cluster endpoint
@@ -458,7 +454,7 @@ func TestDeleteCluster(t *testing.T) {
 	tCtx.CreateRayClusterWithConfigMaps(t, map[string]string{
 		"counter_sample.py": ReadFileAsString(t, "resources/counter_sample.py"),
 		"fail_fast.py":      ReadFileAsString(t, "resources/fail_fast_sample.py"),
-	})
+	}, []rayv1api.RayClusterConditionType{})
 	tests := []GenericEnd2EndTest[*api.DeleteClusterRequest]{
 		{
 			Name: "Delete an existing cluster",
@@ -516,7 +512,7 @@ func TestDeleteCluster(t *testing.T) {
 	}
 }
 
-func createOneClusterInEachNamespaces(t *testing.T, numberOfNamespaces int) []*End2EndTestingContext {
+func createOneClusterInEachNamespaces(t *testing.T, numberOfNamespaces int, expectedConditions []rayv1api.RayClusterConditionType) []*End2EndTestingContext {
 	tCtxs := make([]*End2EndTestingContext, numberOfNamespaces)
 	for i := 0; i < numberOfNamespaces; i++ {
 		tCtx, err := NewEnd2EndTestingContext(t)
@@ -529,7 +525,7 @@ func createOneClusterInEachNamespaces(t *testing.T, numberOfNamespaces int) []*E
 		actualCluster, configMapName := tCtx.CreateRayClusterWithConfigMaps(t, map[string]string{
 			"counter_sample.py": ReadFileAsString(t, "resources/counter_sample.py"),
 			"fail_fast.py":      ReadFileAsString(t, "resources/fail_fast_sample.py"),
-		})
+		}, expectedConditions)
 		t.Cleanup(func() {
 			tCtx.DeleteRayCluster(t, actualCluster.Name)
 			tCtx.DeleteConfigMap(t, configMapName)
@@ -546,7 +542,7 @@ func isMatchingCluster(tCtx *End2EndTestingContext, cluster *api.Cluster) bool {
 // TestGetAllClusters tests gets all Ray clusters from k8s cluster
 func TestGetAllClusters(t *testing.T) {
 	numberOfNamespaces := 3
-	tCtxs := createOneClusterInEachNamespaces(t, numberOfNamespaces)
+	tCtxs := createOneClusterInEachNamespaces(t, numberOfNamespaces, []rayv1api.RayClusterConditionType{})
 	tCtx := tCtxs[0]
 	response, actualRPCStatus, err := tCtx.GetRayAPIServerClient().ListAllClusters(&api.ListAllClustersRequest{})
 	require.NoError(t, err, "No error expected")
@@ -574,7 +570,7 @@ func TestGetAllClusters(t *testing.T) {
 // TestGetAllClustersByPagination tests gets all Ray clusters from k8s cluster with pagination
 func TestGetAllClustersByPagination(t *testing.T) {
 	numberOfNamespaces := 3
-	tCtxs := createOneClusterInEachNamespaces(t, numberOfNamespaces)
+	tCtxs := createOneClusterInEachNamespaces(t, numberOfNamespaces, []rayv1api.RayClusterConditionType{})
 	tCtx := tCtxs[0]
 	gotClusters := make([]bool, numberOfNamespaces)
 	continueToken := ""
@@ -614,7 +610,7 @@ func TestGetAllClustersByPagination(t *testing.T) {
 // TestGetAllClustersByPaginationWithAllResults tests gets all Ray clusters from k8s cluster with pagination returning all results
 func TestGetAllClustersByPaginationWithAllResults(t *testing.T) {
 	numberOfNamespaces := 3
-	tCtxs := createOneClusterInEachNamespaces(t, numberOfNamespaces)
+	tCtxs := createOneClusterInEachNamespaces(t, numberOfNamespaces, []rayv1api.RayClusterConditionType{})
 	tCtx := tCtxs[0]
 	response, actualRPCStatus, err := tCtx.GetRayAPIServerClient().ListAllClusters(&api.ListAllClustersRequest{
 		Limit: 10,
@@ -653,7 +649,7 @@ func TestGetClustersInNamespace(t *testing.T) {
 	cluster, configMapName := tCtx.CreateRayClusterWithConfigMaps(t, map[string]string{
 		"counter_sample.py": ReadFileAsString(t, "resources/counter_sample.py"),
 		"fail_fast.py":      ReadFileAsString(t, "resources/fail_fast_sample.py"),
-	})
+	}, []rayv1api.RayClusterConditionType{})
 	t.Cleanup(func() {
 		tCtx.DeleteRayCluster(t, cluster.Name)
 		tCtx.DeleteConfigMap(t, configMapName)
@@ -690,11 +686,11 @@ func TestGetClustersByPaginationInNamespace(t *testing.T) {
 	cluster1, configMapName1 := tCtx.CreateRayClusterWithConfigMaps(t, map[string]string{
 		"counter_sample.py": ReadFileAsString(t, "resources/counter_sample.py"),
 		"fail_fast.py":      ReadFileAsString(t, "resources/fail_fast_sample.py"),
-	}, "cluster1")
+	}, []rayv1api.RayClusterConditionType{}, "cluster1")
 	cluster2, configMapName2 := tCtx.CreateRayClusterWithConfigMaps(t, map[string]string{
 		"counter_sample.py": ReadFileAsString(t, "resources/counter_sample.py"),
 		"fail_fast.py":      ReadFileAsString(t, "resources/fail_fast_sample.py"),
-	}, "cluster2")
+	}, []rayv1api.RayClusterConditionType{}, "cluster2")
 	t.Cleanup(func() {
 		tCtx.DeleteRayCluster(t, cluster1.Name)
 		tCtx.DeleteRayCluster(t, cluster2.Name)
@@ -733,7 +729,7 @@ func TestGetClustersByNameInNamespace(t *testing.T) {
 	cluster, configMapName := tCtx.CreateRayClusterWithConfigMaps(t, map[string]string{
 		"counter_sample.py": ReadFileAsString(t, "resources/counter_sample.py"),
 		"fail_fast.py":      ReadFileAsString(t, "resources/fail_fast_sample.py"),
-	})
+	}, []rayv1api.RayClusterConditionType{})
 	t.Cleanup(func() {
 		tCtx.DeleteRayCluster(t, cluster.Name)
 		tCtx.DeleteConfigMap(t, configMapName)
@@ -793,33 +789,4 @@ func TestGetClustersByNameInNamespace(t *testing.T) {
 			}
 		})
 	}
-}
-
-func waitForRunningCluster(t *testing.T, tCtx *End2EndTestingContext, clusterName string) {
-	// wait for the cluster to be in a running state for 3 minutes
-	// if is not in that state, return an error
-	err := wait.PollUntilContextTimeout(tCtx.ctx, 500*time.Millisecond, 3*time.Minute, false, func(_ context.Context) (done bool, err error) {
-		rayCluster, err00 := tCtx.GetRayClusterByName(clusterName)
-		if err00 != nil {
-			return true, err00
-		}
-		t.Logf("Found cluster state of '%s' for ray cluster '%s'", rayCluster.Status.State, clusterName)
-		return rayCluster.Status.State == rayv1api.Ready, nil
-	})
-	require.NoErrorf(t, err, "No error expected when getting ray cluster: '%s', err %v", tCtx.GetRayClusterName(), err)
-}
-
-func waitForDeletedCluster(t *testing.T, tCtx *End2EndTestingContext, clusterName string) {
-	// wait for the cluster to be deleted
-	// if is not in that state, return an error
-	err := wait.PollUntilContextTimeout(tCtx.ctx, 500*time.Millisecond, 3*time.Minute, false, func(_ context.Context) (done bool, err error) {
-		rayCluster, err00 := tCtx.GetRayClusterByName(clusterName)
-		if err00 != nil &&
-			assert.EqualError(t, err00, "rayclusters.ray.io \""+tCtx.GetRayClusterName()+"\" not found") {
-			return true, nil
-		}
-		t.Logf("Found status of '%s' for ray cluster '%s'", rayCluster.Status.State, clusterName)
-		return false, err00
-	})
-	require.NoErrorf(t, err, "No error expected when deleting ray cluster: '%s', err %v", clusterName, err)
 }
