+++ apiserver/test/e2e/types.go
@@ -265,7 +265,7 @@ func (e2etc *End2EndTestingContext) DeleteComputeTemplate(t *testing.T) {
 	require.NoErrorf(t, err, "No error expected while deleting a compute template (%s, %s)", e2etc.computeTemplateName, e2etc.namespaceName)
 }
 
-func (e2etc *End2EndTestingContext) CreateRayClusterWithConfigMaps(t *testing.T, configMapValues map[string]string, name ...string) (*api.Cluster, string) {
+func (e2etc *End2EndTestingContext) CreateRayClusterWithConfigMaps(t *testing.T, configMapValues map[string]string, expectedConditions []rayv1api.RayClusterConditionType, name ...string) (*api.Cluster, string) {
 	configMapName := e2etc.CreateConfigMap(t, configMapValues, name...)
 	t.Cleanup(func() {
 		e2etc.DeleteConfigMap(t, configMapName)
@@ -333,17 +333,7 @@ func (e2etc *End2EndTestingContext) CreateRayClusterWithConfigMaps(t *testing.T,
 	})
 	require.NoErrorf(t, err, "No error expected while creating cluster (%s/%s)", e2etc.namespaceName, clusterName)
 
-	// wait for the cluster to be in a running state for 3 minutes
-	// if is not in that state, return an error
-	err = wait.PollUntilContextTimeout(e2etc.ctx, 500*time.Millisecond, 3*time.Minute, false, func(_ context.Context) (done bool, err error) {
-		rayCluster, err00 := e2etc.GetRayClusterByName(actualCluster.Name)
-		if err00 != nil {
-			return true, err00
-		}
-		t.Logf("Found cluster state of '%s' for ray cluster '%s'", rayCluster.Status.State, clusterName)
-		return rayCluster.Status.State == rayv1api.Ready, nil
-	})
-	require.NoErrorf(t, err, "No error expected when getting ray cluster: '%s', err %v", clusterName, err)
+	waitForClusterConditions(t, e2etc, actualCluster.Name, expectedConditions)
 	return actualCluster, configMapName
 }
